requirements.txt
python-dotenv==1.1.0
pydantic==2.10.6
typing_extensions==4.13.2
gspread==6.2.1
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
pydantic_core==2.27.2
livekit==1.0.8
livekit-agents==1.0.23
livekit-api==1.0.2
livekit-plugins-aws==1.0.22
livekit-plugins-deepgram==1.0.22
livekit-plugins-elevenlabs==1.0.22
livekit-plugins-noise-cancellation==0.2.4
livekit-plugins-openai==1.0.23
livekit-plugins-silero==1.0.22
livekit-protocol==1.0.3
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0